dxr: Perform more operations directly against the DB
The indexing phase has been refurbished so more data hits directly
the database and less is retained in memory. The clang plugin has
changed too to adapt to this, resulting in overall lower times for
indexing (from ~3h30m to ~40m) and lower memory usage (peaking at
25% of the original one)

The database schema had to change too in order to allow for perfomant
queries as well for HTML generation, all *loc fields have been
replaced by file_id/line/col triplets, with a foreign key on the
files table, so searching for a given file doesn't result in a full
table scan. With these changes, dxr-index.py is able to generate an
average of 37 HTML files per second.